refactor: MoviesActivity > HomeFragment 변경
