+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,10 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
-            android:exported="false" />
-        <activity
             android:name=".view.reservation.ReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".view.movies.MoviesActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
